requirements.txt
streamlit==1.37.1
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.23
openai==1.43.0
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0
beautifulsoup4==4.12.3
requests==2.32.3
